backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
containers==0.0.4
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dependency-injector==4.49.0
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
filelock==3.29.0
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
gast==0.7.0
gdown==6.0.0
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
kiwisolver==1.5.0
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.6
lightphe==0.0.24
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.33
ml_dtypes==0.5.4
mmh3==5.2.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.1
namex==0.1.0
numpy==1.26.4
onnxruntime-gpu==1.17.0
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
paho-mqtt==2.1.0
pandas==3.0.2
pillow==12.1.1
postgrest==2.28.2
propcache==0.4.1
protobuf==7.34.1
providers==0.0.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyreadline3==3.5.4
pyroaring==1.0.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
realtime==2.28.2
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
scipy==1.17.1
sentry-sdk==2.55.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
soupsieve==2.8.3
starlette==0.52.1
storage3==2.28.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
tf_keras==2.21.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.8
wheel==0.47.0
wrapt==2.1.2
yarl==1.23.0
zlib-state==0.1.12
zstandard==0.25.0
